more ui improvements
