+++ package.json
@@ -14,7 +14,8 @@
     "date-fns": "^3.3.1",
     "lucide-react": "^0.323.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "sonner": "^1.4.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.55",
