+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-
+import { Toaster } from "sonner"
 import { App } from './app'
 import './index.css'
 
@@ -9,13 +9,9 @@ import eruda from 'eruda'
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <App />
+		<Toaster richColors position="bottom-right" />
   </React.StrictMode>,
 )
-
-
-
-
-
 
 if (window) {
 	let el = document.createElement('div')
