+++ src/utils/time.ts
@@ -1,12 +1,12 @@
 /** nLW Expert **/
 
 import { formatDistance } from "date-fns"
-import { ptBR } from "date-fns/locale/pt-BR"
 
 export function getDistance(date:Date) {
-	return formatDistance(date, new Date(), {
-		addSuffix: true,
-		locale: ptBR
-	})
+	return formatDistance(
+		date,
+		new Date(),
+		{ addSuffix: true }
+	)
 }
